refactor(sources): extract shared time-window selector into Source CDK (#956)
Move the relative-duration / absolute-timestamp time-window selector parser out of sources/aws into internal/sourcecdk as a typed ParseTimeSelector helper. Pure refactor with no change to emitted events, attributes, schema refs, URNs, or cursors; aws no-growth ratchet lowered 19121->19070. aws remains the largest grandfathered source and still needs a dedicated multi-PR extraction.
