+++ docs/engineering/source-cdk-extraction.md
@@ -23,7 +23,7 @@ Legacy sources above the 300 LOC budget are grandfathered with exact no-growth c
 | Source | Current LOC Budget | Extraction Pressure |
 | --- | ---: | --- |
 | `aurelius` | 653 | Split source orchestration from record mapping and shared request plumbing. |
-| `aws` | 19121 | Extract common AWS client/session, pagination, ARN parsing, and resource-family traversal helpers. |
+| `aws` | 19070 | Extract common AWS client/session, pagination, ARN parsing, and resource-family traversal helpers. |
 | `azure` | 2856 | Extract subscription traversal, client factories, and paginated resource readers. |
 | `cosmo` | 1104 | Move shared API pagination and response normalization into reusable source helpers. |
 | `gcp` | 2104 | Extract project traversal, service clients, and resource-family pagination helpers. |
