+++ tools/archtests/source_packages_test.go
@@ -15,7 +15,7 @@ const newSourcePackageLOCBudget = 300
 // source grows, move shared behavior into the Source CDK instead of raising it.
 var grandfatheredSourcePackageLOCBudgets = map[string]int{
 	"aurelius":        653,
-	"aws":             19121,
+	"aws":             19070,
 	"azure":           2856,
 	"cosmo":           1104,
 	"gcp":             2104,
